# perf: use `fast-glob` instead of `globby`