--- a/packages/yarnpkg-core/package.json
+++ b/packages/yarnpkg-core/package.json
@@ -24,7 +24,7 @@
     "cross-spawn": "7.0.3",
     "diff": "^5.1.0",
     "dotenv": "^16.3.1",
-    "globby": "^11.0.1",
+    "fast-glob": "^3.2.2",
     "got": "^11.7.0",
     "lodash": "^4.17.15",
     "micromatch": "^4.0.2",